+++ example/packages/edit/src/components/Edit.vue
@@ -1,12 +1,13 @@
 <template>
   <div class="tce-container">
     <div>Times clicked: {{ element.data.count }}</div>
-    <button @click="increment">Increment</button>
+    <button :disabled="isDisabled" @click="increment">Increment</button>
     <div class="background-input-container">
       <label for="backgroundInput">
         Set background:
         <input
           id="backgroundInput"
+          :disabled="isDisabled"
           accept="image/png, image/jpeg"
           type="file"
           @change="upload"
@@ -31,21 +32,25 @@
       width="200px"
     />
     <div class="my-3">
-      <button @click="emit('link')">Link example</button>
+      <button :disabled="isDisabled" @click="emit('link')">Link example</button>
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
-import { defineEmits, defineProps, inject } from 'vue';
 import type { InputFileEvent, StorageApi } from '@tailor-cms/cek-common';
 import { createUploadForm } from '@tailor-cms/cek-common';
 import { Element } from 'tce-manifest';
+import { inject } from 'vue';
 
 const storageService = inject('$storageService') as StorageApi;
 const elementBus = inject('$elementBus') as any;
 
-const props = defineProps<{ element: Element; isFocused: boolean }>();
+const props = defineProps<{
+  element: Element;
+  isFocused: boolean;
+  isDisabled: boolean;
+}>();
 const emit = defineEmits(['save', 'link']);
 
 const increment = () => {
@@ -59,15 +64,13 @@ elementBus.on('decrement', ({ count }: any) => console.log(count));
 const upload = (e: InputFileEvent | any) => {
   const form = createUploadForm(e);
   if (!form) return;
-  return storageService.upload(form).then(({ key, url }) => {
+  return storageService.upload(form).then(({ key, url }) =>
     emit('save', {
       ...props.element.data,
       key,
-      assets: {
-        backgroundUrl: url,
-      },
-    });
-  });
+      assets: { backgroundUrl: url },
+    }),
+  );
 };
 </script>
 
@@ -88,6 +91,12 @@ button {
   padding: 0.125rem 0.625rem;
   background-color: #eee;
   border: 1px solid #444;
+  border-radius: 0.125rem;
+
+  &:disabled {
+    background-color: #eee;
+    border-color: #d6d6d6;
+  }
 }
 
 .upload-details {