+++ packages/runtime/tce-server/common/config.ts
@@ -1,4 +1,4 @@
-import camelCase from 'lodash/camelCase';
+import { camelCase } from 'lodash-es';
 
 export const getTceConfig = (env: object) =>
   Object.keys(env)