+++ packages/runtime/tce-server/content-element/index.ts
@@ -20,7 +20,6 @@ function initRouter({ type, initState, isQuestion, isGradable, hookMap }) {
     hookMap,
   });
 
-  /* eslint-disable @typescript-eslint/no-misused-promises */
   const router = express.Router();
   router.param('id', getContentElement);
   router.route('/').get(getCtrl);