+++ packages/runtime/tce-edit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tailor-cms/tce-edit-runtime
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies.
+
 ## 1.0.4
 
 ### Patch Changes