+++ packages/tce-boot/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tailor-cms/tce-boot
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @tailor-cms/tce-edit-runtime@1.0.5
+  - @tailor-cms/eslint-config@0.0.3,
+
 ## 1.0.7
 
 ### Patch Changes